Add Delete Option for Factors
