+++ core/bulk/bulk.go
@@ -16,6 +16,7 @@ type Bulker interface {
 	BulkInsertDPO(ctx context.Context, requests []core.DPOUpdateRequest) error
 	BulkInsertFactors(ctx context.Context, requests []FactorUpdateRequest) error
 	BulkInsertGlobalFactors(ctx context.Context, requests []GlobalFactorRequest) error
+	BulkDeleteFactor(ctx context.Context, ids []string) error
 }
 
 type Adjuster interface {
